backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
pydantic==2.10.4
google-genai